apps/viz-backend/requirements.txt
pandas>=1.1.3
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic[dotenv]
PyYAML>=5.4.1
requests>=2.25.1
